Move command lint command code into lint folder
